pyproject.toml
[tool.poetry]
name = "ap-faas"
version = "0.1.0"
description = ""
authors = ["Joel Corporan <ap-faas@corporans.com>"]
homepage = "https://github.com/joelcorporan/ap-faas"
repository = "https://github.com/joelcorporan/ap-faas"
license = "MIT"
readme = "README.md"
#packages = [{include = "src/ap_faas"}]

[tool.poetry.scripts]
experiment = "src.ap_faas.app:experiment"
trace = "src.ap_faas.app:trace"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^1.5.3"
aiohttp = "^3.8.4"
asyncio = "^3.4.3"
isort = "^5.12.0"
rich = "^13.3.3"
loguru = "^0.6.0"
jsonschema = "^4.17.3"
numpy = "^1.24.2"
pyyaml = "^6.0"
pandas-stubs = "^1.5.3.230321"
boto3 = "^1.26.109"

[tool.poetry.group.lint.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"

[tool.isort]
profile = "black"

[tool.mypy]
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = false
disallow_subclassing_any = true
# platform
python_version = "3.10"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true
# namespace_packages = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
